refactor: split war_resolver and faction_absorption_engine (#4130)
Move declare-war and offer-peace apply tails into sibling modules and
extract faction absorption apply/cleanup helpers so all diplomacy lib
files stay well under the wave-5 400-line ratchet.

Refs #4130
