+++ packages/colonizethis_diplomacy/lib/src/diplomacy/war_resolver.dart
@@ -1,13 +1,9 @@
-import 'package:colonizethis_data/colonizethis_data.dart';
 import 'package:colonizethis_models/colonizethis_models.dart';
 
-import '../dossier/evidence_rules.dart'
-    show evidenceForDeclareWar, evidenceForOfferPeace;
-import 'diplomacy_relation_updates.dart';
-import 'diplomacy_resolver.dart';
-import 'diplomacy_shared_helpers.dart';
-import 'intervention_resolver.dart';
 import 'diplomacy_event_logging.dart';
+import 'diplomacy_resolver.dart';
+import 'war_resolver_declare_war.dart';
+import 'war_resolver_offer_peace.dart';
 
 Game processWarAndPeace(
   Game game,
@@ -17,7 +13,7 @@ Game processWarAndPeace(
   void Function(DialogueEvent)? onDialogue,
   IntraTurnEventTally? eventTally,
 }) {
-  final peaceOffersByPairKey = _peaceOfferPairKeysForGreatPowers(
+  final peaceOffersByPairKey = peaceOfferPairKeysForGreatPowers(
     game,
     diploByPlayer,
     factionMembership,
@@ -86,29 +82,6 @@ Game _runWarAndPeaceOrders(
   return game;
 }
 
-/// GP–GP peace offers by unordered pair (both sides must offer in same phase).
-Map<String, Set<String>> _peaceOfferPairKeysForGreatPowers(
-  Game game,
-  Map<String, List<DiplomaticOrder>> diploByPlayer,
-  DiplomacyFactionMembership factionMembership,
-) {
-  final peaceOffersByPairKey = <String, Set<String>>{};
-  for (final entry in diploByPlayer.entries) {
-    final gpId = entry.key;
-    for (final order in entry.value) {
-      if (order.type != DiplomaticOrderType.offerPeace) continue;
-      final targetId = order.targetFactionId;
-      if (!factionMembership.isGreatPower(gpId) ||
-          !factionMembership.isGreatPower(targetId)) {
-        continue;
-      }
-      final key = pairKey(gpId, targetId);
-      peaceOffersByPairKey.putIfAbsent(key, () => <String>{}).add(gpId);
-    }
-  }
-  return peaceOffersByPairKey;
-}
-
 Game _applyWarPhaseOrder({
   required Game game,
   required RelationUpsertIndex relationsIndex,
@@ -121,7 +94,7 @@ Game _applyWarPhaseOrder({
   IntraTurnEventTally? eventTally,
 }) {
   if (order.type == DiplomaticOrderType.declareWar) {
-    return _applyDeclareWarOrder(
+    return applyDeclareWarOrder(
       game: game,
       relationsIndex: relationsIndex,
       gpId: gpId,
@@ -132,7 +105,7 @@ Game _applyWarPhaseOrder({
     );
   }
   if (order.type == DiplomaticOrderType.offerPeace) {
-    return _applyOfferPeaceOrder(
+    return applyOfferPeaceOrder(
       game: game,
       relationsIndex: relationsIndex,
       gpId: gpId,
@@ -145,168 +118,4 @@ Game _applyWarPhaseOrder({
     );
   }
   return game;
-}
-
-Game _applyDeclareWarOrder({
-  required Game game,
-  required RelationUpsertIndex relationsIndex,
-  required String gpId,
-  required DiplomaticOrder order,
-  required int turn,
-  void Function(DialogueEvent)? onDialogue,
-  IntraTurnEventTally? eventTally,
-}) {
-  final targetId = order.targetFactionId;
-  final rel = getRelation(game, gpId, targetId);
-  final atPeace = rel == null || rel.atPeace;
-  if (!atPeace) {
-    return game;
-  }
-  if (onDialogue != null && isAiControlledForEvidence(game, gpId)) {
-    onDialogue(
-      DialogueEvent(
-        leaderId: gpId,
-        category: 'diplomatic',
-        situation: 'declare_war',
-        era: 'earlyModern',
-        variables: {'otherNation': targetId},
-      ),
-    );
-  }
-  final evidence = evidenceForDeclareWar(game, gpId, targetId, turn);
-  relationsIndex.upsert(
-    gpId,
-    targetId,
-    warStateRelationUpdater(gpId, targetId, turn),
-  );
-  var nextGame = game.copyWith(
-    diplomacyRelations: committedRelations(relationsIndex),
-    dossierEvidenceEntries: [...game.dossierEvidenceEntries, ...evidence],
-  );
-  nextGame = cancelSubsidiesBetweenGps(
-    nextGame,
-    gpId,
-    targetId,
-    turn,
-    eventTally: eventTally,
-  );
-  nextGame = logDiplomaticEvent(
-    nextGame,
-    turn,
-    DiplomaticEventType.declareWar,
-    {gpId, targetId},
-    fromFactionId: gpId,
-    toFactionId: targetId,
-    wasAiInitiator: isAiControlledForEvidence(nextGame, gpId),
-    eventTally: eventTally,
-    logMessage:
-        'diplomacy war declared $gpId vs $targetId (scores reset to 20)',
-  );
-  return nextGame;
-}
-
-Game _applyOfferPeaceOrder({
-  required Game game,
-  required RelationUpsertIndex relationsIndex,
-  required String gpId,
-  required DiplomaticOrder order,
-  required int turn,
-  required Map<String, Set<String>> peaceOffersByPairKey,
-  required DiplomacyFactionMembership factionMembership,
-  void Function(DialogueEvent)? onDialogue,
-  IntraTurnEventTally? eventTally,
-}) {
-  final targetId = order.targetFactionId;
-  final rel = getRelation(game, gpId, targetId);
-  final key = pairKey(gpId, targetId);
-  final bothGreatPowers =
-      factionMembership.isGreatPower(gpId) &&
-      factionMembership.isGreatPower(targetId);
-  final offerers = peaceOffersByPairKey[key] ?? const <String>{};
-  final collapsedSurvivalPeace = bothGreatPowers &&
-      offerers.length == 1 &&
-      offerers.any(
-        (id) =>
-            oldWorldProvinceCountOwnedBy(game, id) <=
-            kStalledOldWorldProvinceThreshold,
-      );
-  final belowQuotaOutmatchedGpPeace = bothGreatPowers &&
-      offerers.length == 1 &&
-      offerers.any((id) {
-        final own = oldWorldProvinceCountOwnedBy(game, id);
-        final enemyOw = oldWorldProvinceCountOwnedBy(game, targetId);
-        final minLead = own <= kStalledOldWorldProvinceThreshold
-            ? 1
-            : kUnwinnableSoleGpMinProvinceDeficit;
-        return isBelowObserverConquestQuota(own) &&
-            enemyOw >= own + minLead;
-      });
-  final multiFrontConsolidationPeace = bothGreatPowers &&
-      offerers.length == 1 &&
-      offerers.any(
-        (id) => atWarGreatPowerCount(game, id, factionMembership) >= 2,
-      );
-  final soleGpWarConsolidationPeace = bothGreatPowers &&
-      offerers.length == 1 &&
-      offerers.any(
-        (id) => atWarGreatPowerCount(game, id, factionMembership) == 1,
-      );
-  final oneSidedGpPeace = collapsedSurvivalPeace ||
-      belowQuotaOutmatchedGpPeace ||
-      multiFrontConsolidationPeace ||
-      soleGpWarConsolidationPeace;
-  final hasMutualOffer =
-      !bothGreatPowers || offerers.length >= 2 || oneSidedGpPeace;
-  if (rel == null || !rel.atWar) {
-    return game;
-  }
-  var bothSidesAgreed = true;
-  if (factionMembership.isGreatPower(targetId) &&
-      factionMembership.isGreatPower(gpId)) {
-    bothSidesAgreed =
-        (offerers.contains(gpId) && offerers.contains(targetId)) ||
-        oneSidedGpPeace;
-  }
-  if (!bothSidesAgreed) {
-    return game;
-  }
-
-  if (onDialogue != null && isAiControlledForEvidence(game, gpId)) {
-    onDialogue(
-      DialogueEvent(
-        leaderId: gpId,
-        category: 'diplomatic',
-        situation: 'peace_offer',
-        era: 'earlyModern',
-        variables: {'otherNation': targetId},
-      ),
-    );
-  }
-  final evidence = evidenceForOfferPeace(game, gpId, targetId, turn);
-  if (!hasMutualOffer) {
-    return game.copyWith(
-      dossierEvidenceEntries: [...game.dossierEvidenceEntries, ...evidence],
-    );
-  }
-  relationsIndex.upsert(
-    gpId,
-    targetId,
-    peaceRelationUpdater(gpId, targetId, turn),
-  );
-  var nextGame = game.copyWith(
-    diplomacyRelations: committedRelations(relationsIndex),
-    dossierEvidenceEntries: [...game.dossierEvidenceEntries, ...evidence],
-  );
-  nextGame = logDiplomaticEvent(
-    nextGame,
-    turn,
-    DiplomaticEventType.peace,
-    {gpId, targetId},
-    fromFactionId: gpId,
-    toFactionId: targetId,
-    wasAiInitiator: isAiControlledForEvidence(nextGame, gpId),
-    eventTally: eventTally,
-    logMessage: 'diplomacy peace $gpId-$targetId',
-  );
-  return nextGame;
 }
